// graphql-api/src/api/api.service.ts
export type UserInput = {
  id: string;
  userAttributes?: UserAttributesInput | null;
  onboarding?: OnboardingInput | null;
};

export type UserAttributesInput = {
  name?: NameInput | null;
  address?: AddressInput | null;
  phone?: PhoneInput | null;
  dob?: DobInput | null;
  ssn?: SsnInput | null;
};
// ...
export type OnboardingInput = {
  lastActive: number;
  lastComplete: number;
  started?: boolean | null;
  abandoned?: boolean | null;
  display?: Array<OnboardingStepInput | null> | null;
};

export type OnboardingStepInput = {
  id?: number | null;
  active?: boolean | null;
  complete?: boolean | null;
  name?: string | null;
};

export type AgenciesInput = {
  transunion?: TransunionInput | null;
  equifax?: EquifaxInput | null;
  experian?: ExperianInput | null;
};

export type TransunionInput = {
  authenticated?: boolean | null;
  authenticatedOn?: string | null;
  indicativeEnrichmentSuccess?: boolean | null;
  indicativeEnrichmentStatus?: TUStatusRefInput | null;
  getAuthenticationQuestionsSuccess?: boolean | null;
  getAuthenticationQuestionsStatus?: TUStatusRefInput | null;
  verifyAuthenticationQuestionsOTPSuccess?: boolean | null;
  verifyAuthenticationQuestionsOTPStatus?: TUStatusRefInput | null;
  verifyAuthenticationQuestionsKBASuccess?: boolean | null;
  verifyAuthenticationQuestionsKBAStatus?: TUStatusRefInput | null;
  serviceBundleFulfillmentKey?: string | null;
  currentRawQuestions?: string | null;
  currentRawAuthDetails?: string | null;
  authAttempt?: number | null;
  pinRequests?: number | null;
  pinAttempts?: number | null;
  pinCurrentAge?: number | null;
  kbaAttempts?: number | null;
  kbaCurrentAge?: number | null;
  enrollmentKey?: string | null;
  enrollReport?: TUReportResponseInput | null;
  enrollMergeReport?: TUReportResponseInput | null;
  enrollVantageScore?: TUReportResponseInput | null;
  enrolled?: boolean | null;
  enrolledOn?: string | null;
  fulfillReport?: TUReportResponseInput | null;
  fulfillMergeReport?: TUReportResponseInput | null;
  fulfillVantageScore?: TUReportResponseInput | null;
  fulfilledOn?: string | null;
  acknowledgedDisputeTerms?: boolean | null;
  acknowledgedDisputeTermsOn?: string | null;
  disputeServiceBundleFulfillmentKey?: string | null;
  disputeEnrollmentKey?: string | null;
  disputeEnrolled?: boolean | null;
  disputeEnrolledOn?: string | null;
};
// ...
export type EquifaxInput = {
  authenticated?: boolean | null;
};

export type ExperianInput = {
  authenticated?: boolean | null;
};

export type PreferencesInput = {
  showAllAccounts?: ShowAccountsPreferenceInput | null;
};

export type ShowAccountsPreferenceInput = {
  creditCards?: boolean | null;
  collectionsAccounts?: boolean | null;
  installmentLoans?: boolean | null;
  mortgages?: boolean | null;
};

export type DashboardInput = {
  isLoaded?: boolean | null;
  negativeFlagged?: boolean | null;
  negativeCardCount?: number | null;
  negativeCardStatus?: string | null;
  negativeReviewed?: boolean | null;
  negativeStatus?: string | null;
  forbearanceFlagged?: boolean | null;
  forbearanceCardStatus?: string | null;
  forbearanceReviewed?: boolean | null;
  forbearanceStatus?: string | null;
  databreachFlagged?: boolean | null;
  databreachCards?: Array<DataBreachCardInput | null> | null;
  databreachCardStatus?: string | null;
  databreachReviewed?: boolean | null;
  databreachStatus?: string | null;
};

export type DataBreachCardInput = {
  reason?: string | null;
  reviewed?: boolean | null;
  condition?: string | null;
  subscriber?: string | null;
  paragraphs?: Array<string | null> | null;
};

export type NavBarInput = {
  home?: NavBarConfigInput | null;
  report?: NavBarConfigInput | null;
  disputes?: NavBarConfigInput | null;
  settings?: NavBarConfigInput | null;
};

export type NavBarConfigInput = {
  badge?: boolean | null;
};
// ...
export type UpdateAppDataInput = {
  id: string;
  user?: UserInput | null;
  agencies?: AgenciesInput | null;
  preferences?: PreferencesInput | null;
  dashboard?: DashboardInput | null;
  navBar?: NavBarInput | null;
  status?: string | null;
  statusReason?: string | null;
  statusReasonDescription?: string | null;
  lastStatusModifiedOn?: string | null;
  nextStatusModifiedOn?: string | null;
  isLoaded?: boolean | null;
};